feat: fix and expose sparsity
In `pub fn add_reaction` (gillespie.rs:190), the rate was added sparsely but the jump was not.

I added a `sparse: bool` to the Gillespie struct, and add every reaction (rate and jump) as sparse (or not) according to its value.

In the Python code, I exposed it as a keyword-only parameter in the run method, with a default value of False.